requirements.txt
cloudscraper==1.2.71
python-dotenv==1.0.1
Requests==2.32.3
